requirements.txt
streamlit
google-genai
pillow
python-dotenv
# opencv-python
opencv-python-headless